+++ src/schick/seg120.h
@@ -3,7 +3,7 @@ namespace M302de {
 #endif
 
 /* 0x20 */
-signed short init_memory(void);
+signed int init_memory(void);
 /* 0x25 */
 void init_game_state(void);
 /* 0x2a */
@@ -11,7 +11,7 @@ void prepare_dirs(void);
 /* 0x2f */
 void cleanup_game(void);
 /* 0x34 */
-void rabies(struct struct_hero*, signed short);
+void rabies(struct struct_hero*, signed int);
 /* 0x39 */
 /* should be static */
 void init_global_buffer(void);
